Fix zsvjmp for ppc64le.
Summary:

This commit fixes zsvjmp for ppc64le. Before call a __sigsetjmp from a shared
lib we must save LR so siglongjmp can jump in the right place. The problem is
that we must call a shared function using bl instruction that overwrites LR
register. The solution here saves LR in r14 and restore it back when __sigsetjmp
returns as r14 (nonvolatile) is saved on jmpbuf by __sigsetjmp.